# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the API Gateway, BFF (Backend-for-Frontend), and all backend microservices of the ecommerce platform.

    ## Architecture Overview
    The system is composed of the following layers:
    - **API Gateway** (`:8080`) — Single ingress point. Validates JWT access tokens from the `access_token` cookie, enforces role-based access control, and proxies requests to the appropriate downstream service.
    - **BFF Service** (`:8088`) — Aggregates data from multiple services into single, page-optimised responses for the web frontend. All BFF endpoints also flow through the Gateway.
    - **Auth Service** (`:8081`) — Issues and validates JWT access tokens (short-lived) and refresh tokens (long-lived). Both tokens are delivered as HttpOnly cookies.
    - **User Service** (`:8085`) — Manages user profiles and credential changes.
    - **Product Service** (`:8082`) — Product catalogue, category tree, bulk import/export, and presigned S3 upload URLs.
    - **Cart Service** (`:8086`) — Per-user shopping cart and checkout initiation.
    - **Order Service** (`:8083`) — Order lifecycle management.
    - **Payment Service** (`:8087`) — Stripe payment session creation, verification, and webhook ingestion.
    - **Inventory Service** — Stock management, reservation, release, and confirmation as part of the order saga.
...

    ## Authentication
...

    ## Role-Based Access Control
...

...

...

...

...
servers:
...
paths:
...
  /bff/home:
    get:
      tags:
        - BFF
      summary: Home page — aggregated products & categories
      description: |
        Returns a single aggregated payload containing the product list (with default pagination)
        and the full category tree. Designed to populate the storefront home page in a single
        round trip. No authentication required.
      operationId: bffGetHome
      security: []
      responses:
        '200':
          description: Aggregated home page data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BFFHomeResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/profile:
    get:
      tags:
        - BFF
      summary: Profile page — user profile & order history
      description: |
        Returns a combined payload containing the authenticated user's profile information and
        their order history. Requires a valid `access_token` cookie. Designed to populate the
        "My Account" / profile page in a single round trip.
      operationId: bffGetProfile
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Aggregated profile page data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BFFProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/register:
    post:
      tags:
        - BFF
      summary: Register a new user account
      description: |
        Creates a new **user** account (role is always `user` for self-registration).
        On success a verification email is sent to the provided address. The user must
        verify their email before they can log in.

        Returns `409 Conflict` if the email is already registered.

        > **Admin Panel Signup:** To register an admin account directly from the Admin Panel UI, include the `X-Admin-Signup-Secret` header with the secret defined in the backend environment. This bypasses the email verification requirement.

        > **Standard Admin Creation:** Alternatively, use `POST /bff/admin/users` with an existing admin JWT cookie.
      operationId: bffRegister
      security: []
      requestBody:
        $ref: '#/components/requestBodies/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/login:
    post:
      tags:
        - BFF
      summary: Login and receive auth cookies
      description: |
        Authenticates a user with email and password. On success, sets four HttpOnly cookies:
        - `__session` — short-lived JWT access token (15 min). **This is the token the API
          Gateway validates on every subsequent request.**
        - `refresh_token` — long-lived token (7 days) used exclusively with `POST /bff/auth/refresh`.
        - `user_id` — the authenticated user's UUID (readable by JS, used by BFF internally).
        - `user_role` — the authenticated user's role (`user` or `admin`).

        Returns `401 Unauthorized` for invalid credentials or unverified email addresses.
      operationId: bffLogin
      security: []
      requestBody:
        $ref: '#/components/requestBodies/LoginRequest'
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: &authCookieHeader
              description: HttpOnly auth cookies set on successful authentication.
              schema:
                type: string
              example: __session=eyJ...; HttpOnly; Secure; SameSite=None; Path=/
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/logout:
    post:
      tags:
        - BFF
      summary: Logout and clear auth cookies
      description: |
        Invalidates the current session and clears both `access_token` and `refresh_token`
        cookies on the client. Safe to call even if the user is not currently authenticated.
      operationId: bffLogout
      security: []
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/refresh:
    post:
      tags:
        - BFF
      summary: Refresh the access token using the refresh cookie
      description: |
        Issues a new `access_token` cookie using the `refresh_token` cookie. Call this when an
        API request returns `401 Unauthorized` due to an expired access token. Returns `401` if
        the refresh token is missing, expired, or revoked.
      operationId: bffRefreshToken
      security: []
      responses:
        '200':
          description: Generic success message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Operation completed successfully.

          headers:
            Set-Cookie: *authCookieHeader
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/status:
    get:
      tags:
        - BFF
      summary: Check authentication status
      description: |
        Returns whether the current request is authenticated and, if so, the basic identity of
        the authenticated user (ID, email, role). Useful for initialising client-side auth state
        on page load.
      operationId: bffGetAuthStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current authentication status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/verify-email:
    post:
      tags:
        - BFF
      summary: Verify email address with OTP code
      description: |
        Verifies a user's email address using the one-time code sent during registration.
        The `email` and `code` fields are both required. Returns `404` if no pending
        verification exists for the given email.
      operationId: bffVerifyEmail
      security: []
      requestBody:
        $ref: '#/components/requestBodies/VerifyEmailRequest'
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/auth/resend-verification:
    post:
      tags:
        - BFF
      summary: Resend email verification code
      description: |
        Generates a new one-time verification code and resends it to the provided email address.
        Use this when the original code has expired or was not received.
        Returns `404` if the email is not registered, and `400` if the email is already verified.
      operationId: bffResendVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/products:
    get:
      tags:
        - BFF
      summary: List products with filtering, sorting, and pagination
      description: |
        Returns a paginated list of products. Supports filtering by featured status, one or
        more category IDs, and price range. Sort order can be controlled via the `sort`
        parameter. No authentication required — public endpoint.
      operationId: bffListProducts
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/IsFeaturedParam'
        - $ref: '#/components/parameters/CategoryIDsParam'
        - $ref: '#/components/parameters/MinPriceParam'
        - $ref: '#/components/parameters/MaxPriceParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /bff/categories:
    get:
      tags:
        - BFF
      summary: Get the full category tree
      description: |
        Returns the complete category hierarchy as a nested tree. Each node may contain a
        `children` array of sub-categories. No authentication required.
      operationId: bffListCategories
      security: []
      responses:
        '200':
          description: Nested category tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/cart:
    get:
      tags:
        - BFF
      summary: Get the current user's cart
      description: |
        Returns the authenticated user's cart contents, including all items and their quantities.
        Returns an empty `items` array if the cart has not been populated yet.
      operationId: bffGetCart
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /bff/checkout:
    post:
      tags:
        - BFF
      summary: (Deprecated) Initiate checkout — alias for POST /bff/cart/checkout
      description: |
        Convenience alias for `POST /bff/cart/checkout`. Both routes invoke the same checkout
        logic. **Deprecated** — prefer `/bff/cart/checkout` for consistency.
      operationId: bffCheckoutAlias
      deprecated: true
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Checkout initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/orders:
    get:
      tags:
        - BFF
      summary: List the current user's orders
      description: |
        Returns a paginated list of orders belonging to the authenticated user, sorted by
        creation date descending (most recent first).
      operationId: bffListOrders
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /bff/users/profile:
    put:
      tags:
        - BFF
      summary: Update the authenticated user's profile
      description: |
        Updates mutable profile fields (`name`, `phone_number`) for the authenticated user.
        Partial updates are supported — omit fields that should remain unchanged.
      operationId: bffUpdateProfile
      security:
        - cookieAuth: []
      requestBody:
        $ref: '#/components/requestBodies/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /bff/users/change-password:
    post:
      tags:
        - BFF
      summary: Change the authenticated user's password
      description: |
        Changes the password for the currently authenticated user. Requires the current password
        (`old_password`) to be provided as confirmation. All active sessions are invalidated
        after a successful password change. Returns `400` if `old_password` is incorrect.
      operationId: bffChangePassword
      security:
        - cookieAuth: []
      requestBody:
        $ref: '#/components/requestBodies/ChangePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /bff/admin/products:
    get:
      operationId: getBffAdminProducts
      tags: [Admin]
      summary: List products (admin)
      description: Returns a paginated product list from product-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
    post:
      operationId: postBffAdminProducts
      tags: [Admin]
      summary: Create product (admin)
      description: Creates a product in product-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductJSONRequest'
      responses:
        '201': {$ref: '#/components/responses/AdminCreatedResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
  /bff/admin/products/presign:
    get:
      operationId: getBffAdminProductsPresign
      tags: [Admin]
      summary: Get product image presign URL (admin)
      description: |
        Proxies the product-service presign endpoint for new-product image uploads. Use the
        returned `public_url` as the stable URL to persist later on the product record.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - name: sku
          in: query
          required: true
          schema:
            type: string
          example: WIDGET-001
        - name: filename
          in: query
          required: false
          schema:
            type: string
          example: front-view.jpg
        - name: content_type
          in: query
          required: false
          schema:
            type: string
          example: image/jpeg
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
  /bff/admin/categories:
    get:
      operationId: getBffAdminCategories
      tags: [Admin]
      summary: List categories (admin)
      description: Returns categories from product-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
    post:
      operationId: postBffAdminCategories
      tags: [Admin]
      summary: Create category (admin)
      description: Creates category in product-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201': {$ref: '#/components/responses/AdminCreatedResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
  /bff/admin/orders:
    get:
      operationId: getBffAdminOrders
      tags: [Admin]
      summary: List orders (admin)
      description: Returns paginated orders from order-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
  /bff/admin/inventory:
    get:
      tags: [Admin]
      summary: List all inventory items (admin)
      operationId: adminListInventory
      description: |
        Returns a paginated list of all product inventory records from DynamoDB.
        Each record shows available stock, reserved count, and low-stock threshold.
        Requires admin role.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200':
          description: Paginated inventory list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      inventory:
                        type: array
                        items:
                          type: object
                          properties:
                            product_id:
                              type: string
                              format: uuid
                            available:
                              type: integer
                            reserved:
                              type: integer
                            threshold:
                              type: integer
                            updated_at:
                              type: string
                              format: date-time
                      meta:
                        type: object
                        properties:
                          page:
                            type: integer
                          page_size:
                            type: integer
                          count:
                            type: integer
                  error:
                    type: string
              example:
                success: true
                data:
                  inventory:
                    - product_id: 3979556c-2759-4aec-8bb0-5aa5ecde93dc
                      available: 150
                      reserved: 0
                      threshold: 0
                      updated_at: '2026-03-11T00:12:29Z'
                  meta:
                    page: 1
                    page_size: 20
                    count: 5
                error: ''
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
  /bff/admin/inventory/{product_id}:
    put:
      operationId: putBffAdminInventoryByProductId
      tags: [Admin]
      summary: Update inventory (admin)
      description: Updates inventory for a product in inventory-service.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/ProductIDString'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                available:
                  type: integer
                  minimum: 0
                threshold:
                  type: integer
                  minimum: 0
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
  /auth/status:
    get:
      tags:
        - Gateway
        - Auth Service
      summary: Check authentication status
      description: |
        Returns the authenticated identity (ID, email, role) from the current access token
        cookie. Useful for server-side session hydration.
      operationId: authGetStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Authentication status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored as an HttpOnly cookie. Issued by the Auth Service on login or
        token refresh. Short-lived (typically 15 minutes). Pass the `admin` scope value in the
        `security` declaration to indicate an endpoint requires the `admin` role.
...